Adds delegate func, restructures code, fleshes out redeem code
