+++ Sources/SuperwallKit/Models/Web2App/CustomerInfo.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//  SuperwallKit
+//
+//  Created by Yusuf Tör on 14/03/2025.
+//
+
+import Foundation
+
+// TODO: Add docs to all the public types
+public struct CustomerInfo {
+  public let entitlements: [Entitlement]
+  public let redemptions: [RedemptionResult]
+}
